Organized the files and folders
organized the files and folders to make the code more cleaner. added new folders: assets, src and files: ui.py, requirements.py, and also added pictures for the README.MD file (all the pictures are in the assets folder).

Updated README file to provide a clear understanding of the project